FAQ dev command fix

+++ app/projects/FAQ/index.ts
@@ -1,17 +1,17 @@
-import { EmbedBuilder,  Emoji,  GuildEmoji,  Message,  SlashCommandBuilder, SlashCommandSubcommandBuilder } from "discord.js";
+import { EmbedBuilder,  Message,  SlashCommandBuilder, SlashCommandSubcommandBuilder } from "discord.js";
 import { client } from "../../discord";
-import { CANCEL_EMOJI_IDENTIFIER, CANCEL_REACTION_TIMEOUT, EMBED_BACKGROUND, searchFor } from "../../features";
+import { CANCEL_EMOJI_IDENTIFIER, CANCEL_REACTION_TIMEOUT, DEBUG, EMBED_BACKGROUND, GUILD_IDS, searchFor } from "../../features";
 import { BuildEntryFQA, FAQEntry, GET_FQA_ENTIRES, GET_RAW_ENTRIES } from "./load-faq";
-import { Context, RESOURCES, resolveVariables } from "../project-loader";
+import { Context, PRE_LOAD, RESOURCES, resolveVariables } from "../project-loader";
+import { GetInfo } from "../general";
 
-const currentMessages = new Map<string, {message:Message<boolean>, time: NodeJS.Timeout}>();
-
+const currentMessages = new Map<string, {message:Message<boolean>, time: NodeJS.Timeout, userId: string}>();
 client.on("messageReactionAdd", async (e, s)=>{
     if(s.id === e.client.user.id) return;
     const data = currentMessages.get(e.message.id);
     if(data){
-        const {message, time} = data;
-        if(message.id === e.message.id && e.emoji.identifier === CANCEL_EMOJI_IDENTIFIER){
+        const {message, time, userId} = data;
+        if(message.id === e.message.id && s.id === userId && e.emoji.identifier === CANCEL_EMOJI_IDENTIFIER){
             currentMessages.delete(message.id);
             clearTimeout(time);
             await message.delete();
@@ -19,6 +19,8 @@ client.on("messageReactionAdd", async (e, s)=>{
     }
 });
 client.on("messageCreate",async (e)=>{
+    const data = GetInfo()[e.guildId??"null"];
+    if(!(GUILD_IDS.includes(e.guildId??"") && DEBUG) && (!data || !(data.faqChannels[e.channelId]??data.allowFAQ))) return;
     if(e.member?.user.id === e.client.user.id) return;
     const content = e.content;
     if(content.match(/^([ ]+|)\?\?+/g)) {
@@ -35,18 +37,19 @@ client.on("messageCreate",async (e)=>{
             allowedMentions:{parse:[]}
         });
         const timeId = setTimeout(()=>{
-            currentMessages.delete(m.id);
+            if(currentMessages.has(m.id)) currentMessages.delete(m.id);
             m.reactions.removeAll();
         }, CANCEL_REACTION_TIMEOUT);
-        currentMessages.set(m.id, {message:m, time: timeId});
+        currentMessages.set(m.id, {message:m, time: timeId, userId:e.member?.user.id??""});
         m.react(CANCEL_EMOJI_IDENTIFIER);
     } else if(content.match(/^([ ]+|)\!\!+/g) && e.member){
+        return;/*
         const text = content.replaceAll(/^([ ]+|)\!\!+([ ]+|)/g,"").toLowerCase().replaceAll(/[ \-\_\*\/\\\,\;]+/g,"-");
         const a = searchFor(text, ["edit", "create", "edit-tags", "remove"]);
         e.reply({
             flags:[4096],
             content:a +": "+ e.member.roles.cache.map(e=>`<@&${e.id}>`).join(" | ") + " has: Testing role " + e.member.roles.cache.has("1222205753369169920")
-        });
+        });*/
     }
 });
 client.registryCommand(
@@ -76,6 +79,7 @@ client.registryCommand(
         }
     }
 )
+PRE_LOAD.subscribe(()=>currentMessages.clear());
 function buildEmbed(fqa: FAQEntry, context: Context){
     const embed = new EmbedBuilder().setColor(EMBED_BACKGROUND).setTitle("FQA - Title");
     if(fqa.title) embed.setTitle(resolveVariables(fqa.title,context));
